moved uploadFile into its own component since it will be reused.
